+++ intelcamp/configs/configs.json
