+++ intelcamp/models/ffnn.py
