+++ intelcamp/models/rnn.py
