+++ tests/fixtures/v4_exp_dir/error_stats_train.json
@@ -1,0 +1,11 @@
+{
+ "pinball_loss": "0.11399426432820088",
+ "rmse": "20.36745",
+ "nmbe": "0.14463687827179597",
+ "cvrmse": "0.19145546110393097",
+ "gof": "0.16966882468905295",
+ "qs": "6.756907169322174",
+ "ace": "0.9142857142857143",
+ "is": "188.20144957587823",
+ "train_time": "1.093774034"
+}
