+++ tests/fixtures/v5_exp_dir/error_stats_train.json
@@ -1,0 +1,11 @@
+{
+ "pinball_loss": "0.19703279768276102",
+ "rmse": "28.2391393889792",
+ "nmbe": "16.42453744779892",
+ "cvrmse": "2.266626300553023",
+ "gof": "11.723971706732062",
+ "qs": "12.19748770933912",
+ "ace": "1.389621913580247",
+ "is": "527.1795393118732",
+ "train_time": "0.9249121389999999"
+}
